requirements.txt
# =============================================================================
# Dependencias para la Aplicación STREAMLIT de Reportes de Ventas
# =============================================================================
# Instalar con: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Framework Web
# -----------------------------------------------------------------------------
streamlit>=1.28.0

# -----------------------------------------------------------------------------
# Visualización
# -----------------------------------------------------------------------------
plotly>=5.17.0

# -----------------------------------------------------------------------------
# Procesamiento de Datos
# -----------------------------------------------------------------------------
pandas>=2.0.0
numpy>=1.24.0

# -----------------------------------------------------------------------------
# Base de Datos
# -----------------------------------------------------------------------------
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# -----------------------------------------------------------------------------
# HTTP / API
# -----------------------------------------------------------------------------
requests>=2.31.0
aiohttp>=3.9.0

# -----------------------------------------------------------------------------
# Async
# -----------------------------------------------------------------------------
nest_asyncio>=1.5.0

# -----------------------------------------------------------------------------
# Configuración
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# Exportación
# -----------------------------------------------------------------------------
openpyxl>=3.1.0

# -----------------------------------------------------------------------------
# Programación de Tareas
# -----------------------------------------------------------------------------
schedule>=1.2.0

# -----------------------------------------------------------------------------
# Testing (desarrollo)
# -----------------------------------------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# -----------------------------------------------------------------------------
# Calidad de Código (desarrollo)
# -----------------------------------------------------------------------------
# flake8>=6.1.0
# black>=23.9.0
# mypy>=1.5.0
